remove redundant cast

+++ mm/2s2h/BenGui/Menu.cpp
@@ -326,7 +326,7 @@ void Menu::MenuDrawItem(WidgetInfo& widget, uint32_t width, UIWidgets::Colors me
                 if (UIWidgets::Combobox("Renderer API (Needs reload)", &configWindowBackend,
                                         &availableWindowBackendsMap, options)) {
                     Ship::Context::GetRawInstance()->GetConfig()->SetInt("Window.Backend.Id",
-                                                                         (int32_t)(configWindowBackend));
+                                                                        configWindowBackend);
                     Ship::Context::GetRawInstance()->GetConfig()->SetString("Window.Backend.Name",
                                                                             windowBackendsMap.at(configWindowBackend));
                     Ship::Context::GetRawInstance()->GetConfig()->Save();
